+++ src/poetry/core/constraints/generic/union_constraint.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 
-from poetry.core.packages.constraints.base_constraint import BaseConstraint
-from poetry.core.packages.constraints.constraint import Constraint
-from poetry.core.packages.constraints.empty_constraint import EmptyConstraint
-from poetry.core.packages.constraints.multi_constraint import MultiConstraint
+from poetry.core.constraints.generic.base_constraint import BaseConstraint
+from poetry.core.constraints.generic.constraint import Constraint
+from poetry.core.constraints.generic.empty_constraint import EmptyConstraint
+from poetry.core.constraints.generic.multi_constraint import MultiConstraint
 
 
 class UnionConstraint(BaseConstraint):
